refactor: 문서 업로드 커스텀 훅 리팩토링
